tests[Op#57516]: specify `delete_all` usage in custom fields projects delete service
As the `custom_fields_projects` table was not set up conventionally with a primary key, it is not possible to
leverage ActiveRecord#destroy as that's coupled to an id. We therefore use a direct SQL DELETE via `ActiveRecord#delete_all`
which has the disadvantage of skipping any callbacks- including `dependent:` relations.

No `dependent:` rules are strictly relevant in this context (at this time) but should that change, they would either need to be triggered manually
OR this table updated to Rails convention: "avoiding configuration"